Elimina metodo provicional
Elimina metodo provicional y cambio de tipo de variable en la clase Arco.Kilometro

+++ Logica/LogicaGrafo/Arco.cs
@@ -3,7 +3,7 @@
     public class Arco
     {
         public Vertice VerticeDestino { get; set; }
-        public decimal Kilometros { get; set; }
+        public double Kilometros { get; set; }
         public Arco Siguiente { get; set; }
 
         public Arco()
@@ -11,7 +11,7 @@
 
         }
 
-        public Arco(Vertice nDestino, decimal nKm)
+        public Arco(Vertice nDestino, double nKm)
         {
             VerticeDestino = nDestino;
             Kilometros = nKm;
